requirements.txt
cachetools==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-api-python-client==2.96.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-httplib2==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-oauthlib==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.22.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.60.0 ; python_version >= "3.10" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
mysql-connector-python==8.1.0 ; python_version >= "3.10" and python_version < "4.0"
oauth2client==4.1.3 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.21.12 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.10" and python_version < "4.0"
